Various fixes

+++ src/tools/delete-files.ts
@@ -3,29 +3,27 @@ import { promises } from "fs";
 import minimatch from "minimatch";
 import { Episode } from "../lib/episode";
 
-const { rm, readdir, stat } = promises;
-const rootPath = process.argv[2];
-const imageDir = process.argv[3];
+const { rm, readdir } = promises;
 
 const minimatchOpts = { nocase: true };
 
-export async function deleteDir(dir: string) {
-  try {
-    await stat(dir);
-  } catch (e) {
-    if (/ENOENT/.test(e.message)) {
-      console.log(`Directory "${dir}" doesn't exist.`);
-      return;
-    }
-    throw e;
-  }
-  console.log(`Deleting directory ${dir}...`);
-  await rm(`${dir}`, {
-    recursive: true,
-  });
-}
+// export async function deleteDir(dir: string) {
+//   try {
+//     await stat(dir);
+//   } catch (e) {
+//     if (/ENOENT/.test(e.message)) {
+//       console.log(`Directory "${dir}" doesn't exist.`);
+//       return;
+//     }
+//     throw e;
+//   }
+//   console.log(`Deleting directory ${dir}...`);
+//   await rm(`${dir}`, {
+//     recursive: true,
+//   });
+// }
 
-export async function deleteFilesForPattern(dir: string, pat: string) {
+export async function deleteFilesForPattern(dir: string, pat: string, recursive: boolean) {
   const files = (await readdir(dir)) || [];
   const matches = files.filter(file => minimatch(file, pat, minimatchOpts));
   if (!matches.length) {
@@ -33,29 +31,41 @@ export async function deleteFilesForPattern(dir: string, pat: string) {
     return;
   }
 
-  console.log(`Deleting ${matches.length} files matching "${dir}}/${pat}"`);
+  console.log(`Deleting ${matches.length} files matching "${dir}/${pat}"`);
   for (const file of matches) {
-    await rm(`${dir}/${file}`);
+    await rm(`${dir}/${file}`, { recursive });
   }
 }
 
 async function deleteFile(file: string) {
   const dir = path.dirname(file);
   const pat = path.basename(file);
-  await deleteFilesForPattern(dir, pat);
+  await deleteFilesForPattern(dir, pat, false);
+}
+
+async function deleteDir(file: string) {
+  const dir = path.dirname(file);
+  const pat = path.basename(file);
+  await deleteFilesForPattern(dir, pat, true);
 }
 
 export async function deleteProjectFiles(episode: Episode) {
   const workDir = episode.getWorkDir();
   const files = episode.getFileNames();
   const indexFiles = await episode.getDgIndexFiles();
+
+  if (indexFiles.d2vIndex) {
+    await deleteFile(`${workDir}/${indexFiles.d2vIndex}`);
+  }
+  for await (const file of indexFiles.audioStreams) {
+    await deleteFile(`${workDir}/${file}`);
+  }
+
+  await deleteFile(`${workDir}/*.log`);
+  await deleteFile(`${workDir}/*.tmp.*.txt`);
+  await deleteFile(`${workDir}/*${files.timecodeMetrics}`);
+  await deleteFile(`${workDir}/${files.deinterlacedAvi}`);
+  await deleteDir(`${workDir}/${files.rawEncodedFile}`);
+  await deleteFile(`${workDir}/avisynth-tmp-*`);
   await deleteDir(`${workDir}/${files.veaiImageDir}`);
-  await deleteFile(`${workDir}/${indexFiles.d2vIndex}`);
-  throw new Error("not done yet");
-  await deleteFile(`${rootPath}*.log`);
-  await deleteFile(`${rootPath}*.txt`);
-  await deleteFile(`${rootPath}*.d2v`);
-  await deleteFile(`${rootPath}*.m2v`);
-  await deleteFile(`${rootPath}*.mkv`);
-  await deleteDir(imageDir);
 }
